refactor: 장바구니 상품 총합 계산
