+++ src/basic/models/cart.ts
@@ -1,4 +1,4 @@
-import { Product, CartItem } from "../../types";
+import { Product, CartItem, Coupon } from "../../types";
 
 type AddToCartResult =
   | { success: true; cart: CartItem[] }
@@ -117,4 +117,59 @@ export const updateQuantity = (
   };
 };
 
-export const calculateCartTotal = () => {};
+/**
+ * 장바구니 상품 금액 총합을 계산하는 함수
+ */
+export const calculateCartTotal = (cart: CartItem[], coupon: Coupon | null) => {
+  let totalBeforeDiscount = 0;
+  let totalAfterDiscount = 0;
+
+  cart.forEach((item) => {
+    const itemPrice = item.product.price * item.quantity;
+    totalBeforeDiscount += itemPrice;
+    totalAfterDiscount += calculateItemTotal(item);
+  });
+
+  if (coupon) {
+    if (coupon.discountType === "amount") {
+      totalAfterDiscount = Math.max(
+        0,
+        totalAfterDiscount - coupon.discountValue
+      );
+    } else {
+      totalAfterDiscount = Math.round(
+        totalAfterDiscount * (1 - coupon.discountValue / 100)
+      );
+    }
+  }
+
+  return {
+    totalBeforeDiscount: Math.round(totalBeforeDiscount),
+    totalAfterDiscount: Math.round(totalAfterDiscount),
+  };
+};
+
+const getMaxApplicableDiscount = (item: CartItem): number => {
+  const { discounts } = item.product;
+  const { quantity } = item;
+
+  const baseDiscount = discounts.reduce((maxDiscount, discount) => {
+    return quantity >= discount.quantity && discount.rate > maxDiscount
+      ? discount.rate
+      : maxDiscount;
+  }, 0);
+
+  if (item.quantity >= 10) {
+    return Math.min(baseDiscount + 0.05, 0.5); // 대량 구매 시 추가 5% 할인
+  }
+
+  return baseDiscount;
+};
+
+export const calculateItemTotal = (item: CartItem): number => {
+  const { price } = item.product;
+  const { quantity } = item;
+  const discount = getMaxApplicableDiscount(item);
+
+  return Math.round(price * quantity * (1 - discount));
+};
